+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `CSSInliner::options()` that implements the Builder pattern. [#71](https://github.com/Stranger6667/css-inline/issues/71)
 
+### Changed
+
+- Restrict visibility of items in `parser.rs`
+
 ### Performance
 
 - Avoid string allocation in `get_full_url`