+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.5.0] - 2020-08-07
+
 ### Added
 
 - `CSSInliner::options()` that implements the Builder pattern. [#71](https://github.com/Stranger6667/css-inline/issues/71)
@@ -92,7 +94,8 @@
 
 - Initial public release
 
-[Unreleased]: https://github.com/Stranger6667/css-inline/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/Stranger6667/css-inline/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/Stranger6667/css-inline/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/Stranger6667/css-inline/compare/0.3.3...0.4.0
 [0.3.3]: https://github.com/Stranger6667/css-inline/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/Stranger6667/css-inline/compare/0.3.1...0.3.2